permissions for laboratory and imaging

+++ app/Http/Controllers/DatatableController.php
@@ -165,9 +165,11 @@ class DatatableController extends Controller
                                         "<i class='fas fa-user-doctor'></i>" .
                                     "</a>&nbsp;";
                     }
-                    $medical .=   "<a class='btn btn-danger' data-toggle='tooltip' title='Delete' onClick='deleteType($item->id)'>" .
+                    elseif(!in_array(auth()->user()->role, ["Laboratory", "Imaging"])){
+                        $medical .=   "<a class='btn btn-danger' data-toggle='tooltip' title='Delete' onClick='deleteType($item->id)'>" .
                                     "<i class='fas fa-trash'></i>" .
                                 "</a>&nbsp;";
+                    }
                 }
                 else{
                     $medical =  "<a class='btn btn-primary' data-toggle='tooltip' title='Diagnostic Examination' onClick='evaluation($item->user_id)'>" .
